requirements.txt
azure-core==1.32.0
azure-identity==1.19.0
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
idna==3.10
isodate==0.7.2
markdown-it-py==3.0.0
mdurl==0.1.2
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
oauthlib==3.2.2
paramiko==3.5.0
pillow==11.1.0
portalocker==2.10.1
pycparser==2.22
pydo==0.7.0
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.0.1
pywin32==308
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-pixels==3.0.1
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
